# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Lendar API Documentation"
  },
  "servers": [
    {
      "url": "https://lendar-api.herokuapp.com"
    }
  ],
  "paths": {
...
    "/signup/employee": {
      "post": {
        "tags": [
          "signup"
        ],
        "summary": "Create employee user",
        "description": "Creates an account with employee model",
        "operationId": "createUser",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "mobile": {
                  "type": "string"
                },
                "citizen": {
                  "type": "string"
                },
                "education": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "created employee details"
          }
        }
      }
    },
...
    "/signin/employee": {
      "post": {
        "tags": [
          "signin"
        ],
        "summary": "Logs employee into the system",
        "description": "",
        "operationId": "loginEmployee",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "employee token"
          }
        }
      }
    },
...
    "/employee/interested/{id}": {
      "put": {
        "summary": "Add employee to interested list in job post with :id",
        "tags": [
          "employee"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The job id in which employee is interested in",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "interest for job logged"
          }
        }
      }
    }
  }
}